convert zoom to float

+++ api/jeff-weiss-api/api/templates/flexible-image-gallery-template.php
@@ -43,8 +43,8 @@
     return array (
       'module' => 'zoom_image',
       'title' => get_sub_field('zoom_title'),
-      'zoom_max' => (int)get_sub_field('zoom_max'),
-      'zoom_start' => (int)get_sub_field('zoom_start'),
+      'zoom_max' => (float)get_sub_field('zoom_max'),
+      'zoom_start' => (float)get_sub_field('zoom_start'),
       'text_overlay_content' => get_sub_field( 'text_overlay_content'),
       'image' => $image['url'],
       'width' => $image['width'],
